combined: refactor: seperate data managers
- Moved the DatabaseManager and JSONDataManager classes to seperate
files
- Updated the data_access file to import the classes from the new files
